requirements.txt
# ── Data Collection & Processing ──────────────────────────────
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
yfinance>=0.2.50

# ── PDF/HTML Parsing (House & Senate disclosures) ─────────────
pdfplumber==0.11.4
beautifulsoup4==4.12.3
cloudscraper>=1.2.71

# ── ML ────────────────────────────────────────────────────────
scikit-learn==1.5.2
joblib==1.4.2

# ── Database ──────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
pymysql==1.1.1

# ── Backend API ───────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
apscheduler==3.10.4

# ── Gemini ────────────────────────────────────────────────────
google-generativeai==0.8.3

# ── Google Cloud (TTS) ────────────────────────────────────────
google-cloud-texttospeech>=2.17.0

# ── Audio/Video Processing ────────────────────────────────────
ffmpeg-python>=0.2.0

# ── Visualization & Testing ──────────────────────────────────
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.14.1
pytest==8.3.3
